fix gobble not sending events, replace logging
When calling logrus.Fatal(), an syscall.Exit() is triggered, which
immediately terminates the program. Deferred function calls are NOT
invoked in this case.

This is problematic, as sentry-go flushes the cache in a deferred
statement, and in most cases (depending on the scheduling), it has no
chance to finish.

Here, we replace the body of main() with a run() function, which
returns, whether capturing the stack trace and sending it to Sentry
was successful or not. main() will then (after run()'s deferred calls
have finished) either or not call os.Exit() directly.

I've also added/changes a few CLI flags:

- `-dsn` is optional, if SENTRY_DSN is present in the environment
- `-debug` dumps sentry events
- `-env` sets an environment name (defaults to "production")